Create add active child feature
