+++ src/WSRequestHandler.h
@@ -86,7 +86,8 @@ class WSRequestHandler : public QObject {
 		static HandlerResponse HandleReorderSceneItems(WSRequestHandler* req);
 
 		static HandlerResponse HandleAddNewScene(WSRequestHandler * req);
-
+		static HandlerResponse HandleAddActiveChild(WSRequestHandler * req);
+		
 		static HandlerResponse HandleGetStreamingStatus(WSRequestHandler* req);
 		static HandlerResponse HandleStartStopStreaming(WSRequestHandler* req);
 		static HandlerResponse HandleStartStopRecording(WSRequestHandler* req);
