+++ src/WSRequestHandler_Scenes.cpp
@@ -84,7 +84,8 @@ HandlerResponse WSRequestHandler::HandleGetSceneList(WSRequestHandler* req) {
 *
 * @param {String (optional)} `scene` Name of the scene to reorder (defaults to current).
 * @param {Array<Scene>} `items` Ordered list of objects with name and/or id specified. Id preferred due to uniqueness per scene
-* @param {int (optional)} `items[].id` Id of a specific scene item. Unique on a scene by scene basis.
+* @param {int (optional)} `items[].id` Id of a specific scene item. Unique on a scene 
+scene basis.
 * @param {String (optional)} `items[].name` Name of a scene item. Sufficiently unique if no scene items share sources within the scene.
 *
 * @api requests
@@ -140,27 +141,3 @@ HandlerResponse WSRequestHandler::HandleReorderSceneItems(WSRequestHandler* req)
 	return req->SendOKResponse();
 }
 
-
-/**
-* Add New Scene
-*
-* @param {String} `sourceName` Name of the source to which the filter belongs
-* @param {String} `filterName` Name of the filter to reconfigure
-* @param {Object} `filterSettings` New settings. These will be merged to the current filter settings.
-*
-* @api requests
-* @name SetSourceFilterSettings
-* @category sources
-* @since 4.5.0
-*/
-HandlerResponse WSRequestHandler::HandleAddNewScene(WSRequestHandler* req)
-{
-
-	if (!req->hasField("sceneName")) {
-		return req->SendErrorResponse("missing request parameters");
-	}
-	const char* sceneName = obs_data_get_string(req->data, "sceneName");
-	obs_scene_create(sceneName);
-	return req->SendOKResponse();
-
-}
